fix(cli): converge command-surface drift + add drift guard
Three drift fixes in the nebutra CLI surfaced by a governance audit:

- stats.ts: handleStatsOverview/Packages scanned the flat packages/<name>
  layout and found ZERO packages on the categorized monorepo (category
  dirs have no package.json). Add collectPackageDirs() handling both flat
  and packages/<category>/<name> layouts. `nebutra stats` reports real counts.
- metadata.ts (the `nebutra schema` Agent surface): removed the phantom
  `preset` command (never registered) and added the 11 registered commands
  missing (build, lint, typecheck, e2e, workflow, backend, theme, logout,
  upgrade, link, unlink); fixed the stale auth domain.
- services.ts: trimmed 'Services Available' help to real services.

Prevent recurrence: extract buildProgram() into program.ts (single source
of truth) + metadata-surface.test.ts asserts metadata.ts matches the real
registered command set bidirectionally.
